moved to separate feedback-bar module
